# osutil/mount: de-duplicate code to use a list