This is slightly less performant due to the loop and variables, but is more readable and less susceptible to typos and the ilk that plague repetitive code.

For clarity, the difference in performance is:
```
benchmark                                    old ns/op     new ns/op     delta
BenchmarkUnmountFlagsToOptsAllMissing-32     2.90          6.20          +113.79%
BenchmarkUnmountFlagsToOptsMixed-32          48.1          60.6          +25.99%
BenchmarkUnmountFlagsToOptsAllPresent-32     262           285           +8.78%
BenchmarkMountFlagsToOptsAllMissing-32       9.89          9.20          -6.98%
BenchmarkMountFlagsToOptsAllPresent-32       406           474           +16.75%
BenchmarkMountFlagsToOptsMixed-32            66.2          67.0          +1.21%

benchmark                                    old allocs     new allocs     delta
BenchmarkUnmountFlagsToOptsAllMissing-32     0              0              +0.00%
BenchmarkUnmountFlagsToOptsMixed-32          1              1              +0.00%
BenchmarkUnmountFlagsToOptsAllPresent-32     3              3              +0.00%
BenchmarkMountFlagsToOptsAllMissing-32       0              0              +0.00%
BenchmarkMountFlagsToOptsAllPresent-32       4              4              +0.00%
BenchmarkMountFlagsToOptsMixed-32            1              1              +0.00%

benchmark                                    old bytes     new bytes     delta
BenchmarkUnmountFlagsToOptsAllMissing-32     0             0             +0.00%
BenchmarkUnmountFlagsToOptsMixed-32          16            16            +0.00%
BenchmarkUnmountFlagsToOptsAllPresent-32     112           112           +0.00%
BenchmarkMountFlagsToOptsAllMissing-32       0             0             +0.00%
BenchmarkMountFlagsToOptsAllPresent-32       240           240           +0.00%
BenchmarkMountFlagsToOptsMixed-32            16            16            +0.00%
```

This is with go 1.13 and calculated using golang.org/x/tools/cmd/benchcmp